refactor (View) : InputView와 OutputView에 interface 도입
- 테스트 용이성 및 의존성 약화를 위해
